fix: seperate gateway_namespace from namespace
As the higress controller only supports operating fixed namespace/name mcpbridges,
the namespace arg is not enough.
For embedded and external gateway, the namespace will refer to gateway_namespace
arg which defaults to higress-system.
For the incluster gateway, the namespace should be different from gateway_namespace
and it defaults to gpustack-system. We will create ingress for models in namespace
and create mcpbridges registries in gateway_namespace.